refactor: move response_tx types to consensus/tx.h
